pyproject.toml
[project]
name = "small-llm-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "chromadb>=1.3.4",
    "langchain>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-chroma>=0.2.0",
    "langgraph>=0.2.0",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pytest>=8.0.0",
    "ruff>=0.7.0",
]
